refactor: move resources to testclasses in trycatch
