--- src/test/java/spoon/test/trycatch/Main.java
@@ -1,23 +1,0 @@
-package spoon.test.trycatch;
-
-import java.io.BufferedReader;
-import java.io.FileReader;
-
-public class Main {
-
-	public void test() {
-
-		try {
-			System.out.println("test");
-		} catch (Exception e) {
-			e.printStackTrace();
-		}
-
-		try (BufferedReader br = new BufferedReader(new FileReader("test"))) {
-			br.readLine();
-		} catch (Exception e) {
-			e.printStackTrace();
-		}
-	}
-
-}
