+++ src/test/java/spoon/test/trycatch/testclasses/TryCatchResourceClass.java
@@ -1,4 +1,4 @@
-package spoon.test.trycatch;
+package spoon.test.trycatch.testclasses;
 
 import java.io.BufferedReader;
 import java.io.FileReader;
